refactor: Service, Controller 의존관계 변경, 패키지 구성 변경
