+++ lecture-service/src/main/java/io/tdd/lectureservice/lecture/domain/LectureService.java
@@ -1,5 +1,6 @@
 package io.tdd.lectureservice.lecture.domain;
 
+import io.tdd.lectureservice.lecture.interfaces.LectureRepository;
 import io.tdd.lectureservice.lecture.interfaces.dto.LectureResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -7,7 +8,6 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -23,11 +23,11 @@ public class LectureService {
      * @param date 조회할 날짜
      * @return lecture dto list
      */
-    public List<LectureResponse> search(LocalDate date) {
+    public List<Lecture> search(LocalDate date) {
         // 주어진 날짜를 해당 날짜의 시작 시간과 끝 시간 설정
         LocalDateTime startDate = date.now().atStartOfDay();
         LocalDateTime endDate = date.now().atTime(LocalTime.MAX);
-        return lectureRepository.findAllByLectureDateBetween(startDate, endDate).stream().map(LectureResponse::new).collect(Collectors.toList());
+        return lectureRepository.findAllByLectureDateBetween(startDate, endDate);
     }
 
 }
