+++ lecture-service/src/main/java/io/tdd/lectureservice/user/domain/UserService.java
@@ -3,6 +3,7 @@ package io.tdd.lectureservice.user.domain;
 import io.tdd.lectureservice.support.common.validation.Validation;
 import io.tdd.lectureservice.support.exception.CustomNotFoundException;
 import io.tdd.lectureservice.support.exception.InvalidException;
+import io.tdd.lectureservice.user.interfaces.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
