+++ lecture-service/src/main/java/io/tdd/lectureservice/user/interfaces/UserRepository.java
@@ -1,9 +1,8 @@
-package io.tdd.lectureservice.user.domain;
+package io.tdd.lectureservice.user.interfaces;
 
+import io.tdd.lectureservice.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
